% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dist.meas}
\alias{dist.meas}
\title{Compute the Distance in Measure between the clustering induced by a kernel density estimator (based on a sample x and a bandwidth h) and the population clustering defined by a K-component normal mixture density.}
\usage{
dist.meas(x, h, mus = 0, sigmas = 1, props = 1, plot = FALSE)
}
\arguments{
\item{x}{(vector) the data to be partitioned.}

\item{h}{the bandwidth to be used to estimate the density via KDE.}

\item{mus}{vector of means of the mixture components.}

\item{sigmas}{vector of standard deviations of the mixture components.}

\item{props}{vector of mixing proportions of the mixture components.}

\item{plot}{if true, the true density and the estimated one are displayed.}
}
\value{
the value of the Distance in Measure.
}
\description{
Compute the Distance in Measure between the clustering induced by a kernel density estimator (based on a sample x and a bandwidth h) and the population clustering defined by a K-component normal mixture density.
}
\references{
Chacón, J.E. (2015). A population background for nonparametric density-based clustering. Statistical Science 30(4): 518-532.
}
