% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Edist.meas_singleh}
\alias{Edist.meas_singleh}
\title{Compute the Expected Distance in Measure (EDM) between a kernel estimator-induced partition and the population one defined by a K-component normal mixture density for a single bandwidth value.}
\usage{
Edist.meas_singleh(n = 100, h, mus = 0, sigmas = 1, props = 1,
  B = 100)
}
\arguments{
\item{n}{sample size of the simulated Monte Carlo samples.}

\item{h}{value of the bandwidth for which the EDM has to be computed.}

\item{mus}{vector of means of the mixture components.}

\item{sigmas}{vector of standard deviations of the mixture components.}

\item{props}{vector of mixing proportions of the mixture components.}

\item{B}{the number of Monte Carlo samples.}
}
\value{
the value of the EDM for the given value \verb{h}
}
\description{
Compute the Expected Distance in Measure (EDM) between a kernel estimator-induced partition and the population one defined by a K-component normal mixture density for a single bandwidth value.
}
\references{
Chacón, J.E. (2015). A population background for nonparametric density-based clustering. Statistical Science 30(4): 518-532.
}
