% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hab1}
\alias{hab1}
\title{Bandwidth \verb{h_AB1} (3.6)}
\usage{
hab1(data)
}
\arguments{
\item{data}{data}
}
\value{
a value for the modal clustering-oriented optimal bandwidth
}
\description{
Compute the optimal bandwidth \verb{h_AB1} (3.6) in Casa et al. (2019).
}
\references{
Casa A., Chacón, J.E. and Menardi, G. (2019). Modal clustering asymptotics with applications to bandwidth selection (https://arxiv.org/pdf/1901.07300.pdf).
}
