% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hab2}
\alias{hab2}
\title{Bandwidth \verb{h_AB2} (3.7)}
\usage{
hab2(data, n)
}
\arguments{
\item{data}{data}

\item{n}{the sample size}
}
\value{
a value for the modal clustering-oriented optimal bandwidth
}
\description{
Compute the optimal bandwidth \verb{h_AB2} (3.7) in Casa et al. (2019).
}
\references{
Casa A., Chacón, J.E. and Menardi, G. (2019). Modal clustering asymptotics with applications to bandwidth selection (https://arxiv.org/pdf/1901.07300.pdf).
}
